Add config for server timeout
